controllers

+++ feelio-backend/src/models/PostModel.js
@@ -11,10 +11,22 @@ class PostModel {
     );
   }
 
+  /*
   findByUser({ user_id, cb }) {
     DBService.dbPool.query(
       "SELECT body, image, user_id, posted_at from posts where user_id = ? ORDER BY post_id DESC",
       [user_id],
+      (error, results) => {
+        cb(error, results);
+      }
+    );
+  }
+  */
+
+  findByFriends({ user_id, cb }) {
+    DBService.dbPool.query(
+      "SELECT body, image, user_id, posted_at from posts where user_id IN (SELECT IF(sender_id = ?, receiver_id, sender_id) as user_id FROM friend_requests WHERE status = 'Accepted' AND (sender_id = ? OR receiver_id = ?)) ORDER BY post_id DESC",
+      [user_id, user_id, user_id],
       (error, results) => {
         cb(error, results);
       }
@@ -31,10 +43,10 @@ class PostModel {
     );
   }
 
-  edit({ body, image, post_id, cb }) {
+  edit({ user_id, body, image, post_id, cb }) {
     DBService.dbPool.query(
-      "UPDATE posts SET body = ?, image = ? WHERE post_id = ?",
-      [body, image, post_id],
+      "UPDATE posts SET body = ?, image = ? WHERE post_id = ? AND user_id = ?",
+      [body, image, post_id, user_id],
       (error, results) => {
         cb(error, results.insertId);
       }
@@ -42,10 +54,10 @@ class PostModel {
   }
 
 
-  delete({ post_id, cb }) {
+  delete({ user_id, post_id, cb }) {
     DBService.dbPool.query(
-      "DELETE FROM posts WHERE post_id = ?",
-      [post_id],
+      "DELETE FROM posts WHERE post_id = ? AND user_id = ?",
+      [post_id, user_id],
       (error, results) => {
         cb(error, results.insertId);
       }
